+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Enhancements
   * [Docker] Add command `azk docker`
+  * [Docker] Support container annotations.
   * [Vm] Adding ssh keys generator, and uploads the key to start vm.
   * [Generators] Now Generators and rules is a extension of the UI.
   * [Manifest] Validate: image is required.